Rewrite the text line manager and garbage collector.
No longer a hard limit on the number of lines in a file,
or the current collection of files,
or all the files you have seen in your edbrowse session.
Still more work to do regarding malloc failures and integer index overflow.

+++ src/url.c
@@ -545,7 +545,7 @@ static void squashDirectories(char *url)
 	char *s, *t, *end;
 	char *inPath = NULL;
 	char *outPath;
-	size_t outPathLen = 0;
+	int outPathLen = 0;
 	char *rest = NULL;
 
 	outPath = initString(&outPathLen);
